Set default sub-market and symbol to first child on dropdown
for trade definition
